Handle empty argument for mac pruning (#338)
Also add MDNS v4 packet to default pruning

(cherry picked from commit 5109a909aea3a21cd3be548d01c3f7148a95381e)

+++ agent-ovs/ovs/OVSRenderer.cpp
@@ -604,7 +604,7 @@ static void convertPruneFilter(std::shared_ptr<PacketDropLogPruneSpec> &sourceSp
     }
     if(sourceSpec->ipProto) {
         std::stringstream strIpProto;
-        strIpProto << sourceSpec->ipProto.get();
+        strIpProto << (uint16_t)sourceSpec->ipProto.get();
         filter->setField(TFLD_IP_PROTO,strIpProto.str());
     }
     if(sourceSpec->sport) {
